refactor(json): isolate parse reuse state
