+++ scripts/shape_descriptor_census_baseline.json
@@ -37,6 +37,11 @@
     "crates/perry-runtime/src/json/mod.rs|keys_array|declaration|keys_array: arr,": 1,
     "crates/perry-runtime/src/json/mod.rs|keys_array|declaration|pub(crate) keys_array: *mut crate::ArrayHeader,": 1,
     "crates/perry-runtime/src/json/parse_inline_object.rs|keys_array|access|(entry.keys_array, entry.shape_id)": 1,
+    "crates/perry-runtime/src/json/parse_reuse.rs|keys_array|access|crate::gc::runtime_write_barrier_root_raw_ptr(entry.keys_array);": 1,
+    "crates/perry-runtime/src/json/parse_reuse.rs|keys_array|access|entry.keys_array,": 1,
+    "crates/perry-runtime/src/json/parse_reuse.rs|keys_array|access|visitor.visit_raw_mut_ptr_slot(&mut entry.keys_array);": 1,
+    "crates/perry-runtime/src/json/parse_reuse.rs|keys_array|declaration|keys_array: *mut crate::array::ArrayHeader,": 1,
+    "crates/perry-runtime/src/json/parse_reuse.rs|keys_array|declaration|keys_array: crate::object::object_keys_array(object),": 1,
     "crates/perry-runtime/src/json/parser.rs|keys_array|access|parser.hot_shape_array = entry.keys_array;": 1,
     "crates/perry-runtime/src/json/parser.rs|keys_array|access|shape.keys_array,": 1,
     "crates/perry-runtime/src/json/parser.rs|keys_array|declaration|pub(crate) keys_array: *mut crate::array::ArrayHeader,": 1,
@@ -57,9 +62,9 @@
   },
   "summary": {
     "codegen_object_header_size_sites": 43,
-    "raw_member_files": 11,
+    "raw_member_files": 12,
     "raw_member_sites": {
-      "keys_array": 28
+      "keys_array": 33
     }
   }
 }
